backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloud-sql-python-connector==1.18.3
cryptography==45.0.5
dnspython==2.7.0
Flask==3.1.1
frozenlist==1.7.0
google-auth==2.40.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
